+++ Common/src/main/java/com/tristankechlo/healthcommand/HealthCommandMain.java
@@ -1,0 +1,17 @@
+package com.tristankechlo.healthcommand;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+public final class HealthCommandMain {
+
+    public static final String MOD_NAME = "HealthCommand";
+    public static final Logger LOGGER = LogManager.getLogger(MOD_NAME);
+    public static final String MOD_ID = "healthcommand";
+    public static final String GITHUB_URL = "https://github.com/tristankechlo/HealthCommand";
+    public static final String GITHUB_ISSUE_URL = GITHUB_URL + "/issues";
+    public static final String DISCORD_URL = "https://discord.gg/bhUaWhq";
+    public static final String CURSEFORGE_URL = "https://curseforge.com/minecraft/mc-mods/health-command";
+    public static final String MODRINTH_URL = "https://modrinth.com/mod/health-command";
+
+}