+++ Common/src/main/java/com/tristankechlo/healthcommand/config/values/IConfigValue.java
@@ -1,0 +1,17 @@
+package com.tristankechlo.healthcommand.config.values;
+
+import com.google.gson.JsonObject;
+
+public interface IConfigValue<T> {
+
+    String getIdentifier();
+
+    void setToDefault();
+
+    T get();
+
+    void serialize(JsonObject jsonObject);
+
+    void deserialize(JsonObject jsonObject);
+
+}