+++ Changelog.md
@@ -1,5 +1,20 @@
 # Changelog
 
+### Version 1.18.2 - 1.3
+ - port to 1.18.2
+ - move config to json format to have the same config on forge and fabric
+ - add command `healthcommand`
+    - `/healthcommand github` shows the url to the github page
+    - `/healthcommand discord` shows the url to the discord server
+    - `/healthcommand wiki` shows the url to the wiki
+    - `/healthcommand issues` shows the url to the issues page
+    - `/healthcommand curseforge` shows the url to the curseforge page
+    - `/healthcommand modrinth` shows the url to the modrinth page
+    - `/healthcommand config`
+       - `healthcommand config reload` reloads the config
+       - `healthcommand config show` shows the config
+       - `healthcommand config reset` resets the config to its default values
+
 ### Version 1.18.1 - 1.2
  - port to 1.18.1